Create a /Checkout folder in /pages, and move /CheckoutDetails and /CheckoutPayment folders there
